Ask the device that is about to open, not the default one
preferredSampleRate() took no argument and answered for the system default,
with a comment beside it reading "since nothing selects another one yet" -- a
statement that stopped being true when the output device picker landed and was
never revisited. It takes a device id now, and AudioEngine::play() passes
chosenDeviceId(), which it already had two lines further down.

It only bites where the track's own rate was refused, which today means DSD: a
listener with a DAC selected in preferences got the *default* device's mix rate
negotiated, and the track resampled to it, while an entirely different device
was opened. Everything else about that playback is correct, which is why it
could sit there.

A parameter rather than remembered state, because the question is asked before
start() and there is nothing to remember yet. An id that no longer resolves
falls through to the default device on purpose -- start() does the same with
one, and the two disagreeing would mean negotiating a rate for a device that
was never going to be opened. The parameter is defaulted, so OfflineOutput and
the test doubles that never overrode it are untouched.

Two cases in test_output_device.cpp, and the second is the one that matters:
an empty selection must stay empty, since an implementation that always passed
something would satisfy the first on its own. Both were run against the
unfixed engine to confirm they fail -- "" == "offline" -- rather than assumed
to. What they do not cover is MiniaudioOutput asking the right device, which
needs a machine with two and someone listening; the wiring is what is pinned,
and the wiring is where the bug was.

Refusing 44,100 in the fake output rather than the DSD rates that provoke this
in earnest is what keeps the cases runnable anywhere. The question is which
device the engine names, and that is the same question whatever rate raised it.

Also the one real warning in the tree: C4267 in test_realfft.cpp, where
`1234U + size` is a size_t handed to an unsigned seed parameter. The seed only
has to differ per size, so the cast says so.

525 cases green, no warnings on a full build of every target.

+++ tests/core/test_realfft.cpp
@@ -77,7 +77,10 @@ TEST_CASE("the real FFT agrees with a direct DFT on random input", "[audio][fft]
     // at N/4 -- a size where that midpoint lands differently is a different code
     // path through the same loop.
     for (const std::size_t size : {4U, 8U, 64U, 256U, 1024U}) {
-        const std::vector<double> input = randomSignal(size, 1234U + size);
+        // The seed only has to differ per size; the cast is what says so, since
+        // `1234U + size` is a size_t and the parameter is not.
+        const std::vector<double> input =
+            randomSignal(size, static_cast<unsigned>(1234U + size));
 
         const RealFft       fft(size);
         std::vector<double> real(fft.bins());
